Minor style changes
